Update to python exercise set.
I have made the addition of a problem between the original problems 27 and 28. This new problem requires a function to be made that sums two matrices created with NumPy arrays. I believe the explanation is more than sufficient to complete the task, but I can elaborate more if need be.